migrated NextJS pages to new app directory
#55
