+++ app/kontakt/page.tsx
@@ -1,4 +1,4 @@
-import { ContactForm } from "../components/ContactForm";
+import { ContactForm } from "./ContactForm";
 
 export default function KontaktPage() {
   return (
