+++ app/not-found.tsx
@@ -1,4 +1,4 @@
-export default function Custom404() {
+export default function NotFound() {
   return (
     <main className="default-padding mx-auto max-w-3xl grow bg-stone-50">
       404 - Sorry bro!
