+++ app/veranstaltungen/VeranstaltungCard.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { formatDate } from "../utils/formatDate";
+import { formatDate } from "../../utils/formatDate";
 
 interface Props {
   titel: string;
