+++ package.json
@@ -18,10 +18,9 @@
     "@types/react-dom": "18.0.10",
     "daisyui": "^2.51.5",
     "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
-    "next": "13.1.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.4.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.5",
     "sanitize-html": "^2.9.0",
     "typescript": "4.9.4",
@@ -33,6 +32,7 @@
     "@types/sanitize-html": "^2.8.0",
     "autoprefixer": "^10.4.13",
     "cypress": "^12.9.0",
+    "eslint-config-next": "^13.4.9",
     "eslint-plugin-tailwindcss": "^3.12.0",
     "postcss": "^8.4.20",
     "prettier": "^2.8.7",
