--- pages/neuigkeiten/index.tsx
@@ -1,147 +1,0 @@
-import Head from "next/head";
-import { getNeuigkeiten } from "../../api/api";
-import { Neuigkeit } from "../../model/Neuigkeit";
-import { Button } from "../../components/Button";
-import { useEffect, useState } from "react";
-import { GetStaticProps, InferGetStaticPropsType } from "next";
-import Link from "next/link";
-import Image from "next/image";
-import { formatDate } from "../../utils/formatDate";
-
-export const getStaticProps: GetStaticProps<{
-  neuigkeiten: Array<Neuigkeit>;
-  totalCount: number;
-}> = async () => {
-  const { neuigkeiten, pagination } = await getNeuigkeiten(6);
-  return {
-    props: { neuigkeiten: neuigkeiten, totalCount: pagination.total },
-  };
-};
-
-export default function NeuigkeitenPage({
-  neuigkeiten,
-  totalCount,
-}: InferGetStaticPropsType<typeof getStaticProps>) {
-  const [displayedNeuigkeiten, setDisplayedNeuigkeiten] =
-    useState<Array<Neuigkeit>>(neuigkeiten);
-  const [page, setPage] = useState(1);
-  const [isLoading, setIsLoading] = useState(false);
-
-  async function getMoreNeuigkeiten() {
-    setIsLoading(true);
-    const nextPage = page + 1;
-    const { neuigkeiten } = await getNeuigkeiten(6, nextPage);
-    setDisplayedNeuigkeiten((prevNeuigkeiten) => {
-      const updatedNeuigkeiten = [...prevNeuigkeiten, ...neuigkeiten];
-      sessionStorage.setItem("neuigkeiten", JSON.stringify(updatedNeuigkeiten));
-      return updatedNeuigkeiten;
-    });
-    setPage(nextPage);
-    sessionStorage.setItem("page", nextPage.toString());
-    setIsLoading(false);
-  }
-
-  useEffect(() => {
-    const position = sessionStorage.getItem("position");
-    const shouldRestorePreviousScrollPosition = sessionStorage
-      .getItem("route")
-      ?.startsWith("/neuigkeiten/");
-    if (position !== null && shouldRestorePreviousScrollPosition) {
-      window.scrollTo(0, parseInt(position));
-    }
-    setPage(parseInt(sessionStorage.getItem("page") ?? "1"));
-    const neuigkeitenInStorage = sessionStorage.getItem("neuigkeiten");
-    if (neuigkeitenInStorage !== null) {
-      setDisplayedNeuigkeiten(JSON.parse(neuigkeitenInStorage));
-    }
-  }, []);
-
-  return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className="default-padding flex flex-col gap-6 bg-base-950 py-12">
-        <div className="group/container mx-auto grid max-w-screen-lg justify-center gap-6 sm:grid-cols-2 lg:grid-cols-3">
-          {displayedNeuigkeiten.map((neuigkeit) => (
-            <PostCard
-              slug={neuigkeit.attributes.slug}
-              key={neuigkeit.attributes.slug}
-              titel={neuigkeit.attributes.titel}
-              vorschautext={neuigkeit.attributes.vorschautext}
-              datum={neuigkeit.attributes.datum}
-              vorschaubild={
-                neuigkeit.attributes.vorschaubild.data.attributes.url
-              }
-            />
-          ))}
-        </div>
-        {displayedNeuigkeiten.length < totalCount && (
-          <div className="mx-auto w-fit">
-            <Button
-              onClick={getMoreNeuigkeiten}
-              disabled={isLoading}
-              className={isLoading ? "loading" : undefined}
-            >
-              Mehr anzeigen
-            </Button>
-          </div>
-        )}
-      </main>
-    </>
-  );
-}
-
-interface Props {
-  titel: string;
-  vorschautext: string;
-  datum: string;
-  vorschaubild: string;
-  slug: string;
-}
-
-const PostCard = ({
-  titel,
-  vorschaubild,
-  vorschautext,
-  datum,
-  slug,
-}: Props) => (
-  <Link
-    href={`/neuigkeiten/${slug}`}
-    key={slug}
-    className="group rounded-xl transition-all duration-500 hover:!opacity-100 group-hover/container:opacity-50"
-    onClick={() =>
-      sessionStorage.setItem("position", window.scrollY.toString())
-    }
-  >
-    <div className="relative h-[24rem] w-full max-w-md overflow-hidden rounded-xl transition-all">
-      <div className="absolute inset-0 h-full shrink-0">
-        <Image
-          src={vorschaubild}
-          alt=""
-          fill
-          className="rounded-xl object-cover object-top transition-all duration-500 group-hover:scale-105"
-          sizes="(max-width: 712px) 100vw, (max-width: 1072px) 50vw, 33vw"
-          priority
-        />
-      </div>
-      <article className="relative z-10 h-full rounded-md bg-gradient-to-b from-transparent to-base-900 p-6">
-        <div className="relative top-16 flex h-full flex-col justify-end gap-3 transition-all duration-500 group-hover:top-0">
-          <time dateTime={datum} className="text-extrasmall text-base-300">
-            {formatDate(new Date(datum))}
-          </time>
-          <h2 className="text-large line-clamp-3 max-w-xs font-semibold text-base-200">
-            {titel}
-          </h2>
-          <p className="line-clamp-2 text-transparent transition-all duration-500 group-hover:text-base-300">
-            {vorschautext}
-          </p>
-        </div>
-      </article>
-    </div>
-  </Link>
-);
