+++ tailwind.config.js
@@ -2,6 +2,7 @@
 const colors = require("tailwindcss/colors");
 module.exports = {
   content: [
+    "./app/**/*.{js,ts,jsx,tsx}",
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
   ],
